stores/requirements.txt
altgraph==0.17.4
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
boto3==1.35.5
botocore==1.35.5
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.41.0
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==42.0.8
daphne==4.1.2
dj-database-url==2.2.0
Django==5.0.6
django-allauth==65.3.0
django-celery-beat==2.6.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
django-hosts==6.0
django-htmx==1.21.0
django-organizations==2.5.0
django-sms==0.7.0
django-tenants==3.7.0
django-timezone-field==7.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
elevenlabs==1.7.0
et_xmlfile==2.0.0
gTTS==2.5.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
jmespath==1.0.1
kombu==5.4.0
msgpack==1.0.8
numpy==2.2.1
openpyxl==3.1.5
packaging==24.0
pandas==2.2.3
pillow==10.3.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
PyJWT==2.8.0
pyOpenSSL==24.1.0
PyPDF2==3.0.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.91
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
requests==2.32.3
s3transfer==0.10.2
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
websockets==13.0
whitenoise==6.6.0
zope.interface==6.4.post2
